Move() frontend complete
Massive overhaul in the workflow of my codebase:

In order to move your actors, you must call the static ActorHeader.Move(...) func by passing in your actor and the timespace delta you wish to operate on.

This is a huge step forward in my opinion, as it marks the nearby end of my 2+ year quest to write a competent  character controller in Unity.

I also like the new workflow design involving a single header file operating on chunks of data (Actor passed in -> ActorHeader -> Actor changed as a result) kinda philosophy in design for this project.

Callbacks and behavioural effects are on the todo list, as well as platforms and physics-interaction integration.

I hope you find this package useful :)