Concurrent State Execution Implementation
I've implemented an experimental new form of a character state machine in which states are appended using a Monobehaviour Wrapper, and subscribed to a machine that uses a SortedList, and SortedDictionary to store them in two databases:
The first database is to have them registered and understood by the machine.

The second database is to sequentially run through a SortedList of executions to be ran in a specified order.

As well as this, I'm planning on using this for the Actor to have an easier time writing up state functionality for the player.